feat: add lookups-related models in appsutils/models/lookups.py
