+++ project/appsutils/models/__init__.py
@@ -1,0 +1,5 @@
+# Django requires registering the models in the folder to see/use them
+from .base_entity import BaseEntity
+from .enums import Gender, RecordStatus
+from .lookups import LookUp, LookUpCategories
+
