+++ project/appsutils/models/base_entity.py
@@ -1,39 +1,11 @@
-import enum
-
 from django.db import models
 from django.conf import settings
 from django.utils import timezone
+from .enums import RecordStatus
 
 # Create your models here.
 
 User = settings.AUTH_USER_MODEL
-
-
-class RecordStatus(enum.Enum):
-    """
-    This enum class represents the status of every record in the database.
-    The status can be ACTIVE, DELETED, ACTIVE_LOCKED.
-    The default record status for new database entries is ACTIVE.
-    """
-    ACTIVE = "Active"
-    DELETED = "Delete"
-    ACTIVE_LOCKED = "Active Locked"
-
-    def __str__(self):
-        return self.name
-
-
-class Gender(enum.Enum):
-    """
-    This enum class represents the gender of users.
-    Can be Male, Female, Other.
-    """
-    MALE = "Male"
-    FEMALE = "Female"
-    OTHER = "Other"
-
-    def __str__(self):
-        return self.name
 
 
 class BaseEntity(models.Model):
